+++ src/main/kotlin/lotto/domain/LottoConstants.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 object LottoConstants {
