+++ src/main/kotlin/lotto/domain/LottoRandomGenerator.kt
@@ -1,8 +1,8 @@
-package lotto
+package lotto.domain
 
-import lotto.LottoConstants.LOTTO_SIZE
-import lotto.LottoConstants.MAX_LOTTO_NUMBER
-import lotto.LottoConstants.MIN_LOTTO_NUMBER
+import lotto.domain.LottoConstants.LOTTO_SIZE
+import lotto.domain.LottoConstants.MAX_LOTTO_NUMBER
+import lotto.domain.LottoConstants.MIN_LOTTO_NUMBER
 
 object LottoRandomGenerator {
     fun randomGenerate(): Lotto {
