+++ src/test/kotlin/lotto/WinningResultTest.kt
@@ -1,6 +1,6 @@
 package lotto
 
-import WinningResult
+import lotto.domain.WinningResult
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
